backend/requirements_cpu.txt
fastapi
uvicorn
torch
transformers
psycopg2-binary
requests
sqlalchemy
uvloop
passlib[bcrypt]
email-validator
python-jose[cryptography]
httpx


python-multipart
sentencepiece==0.1.99
# Document parsing

python-pptx==0.6.21  # For PPT extraction
# Optional PostgreSQL extensions
pgvector==0.2.1  # For more efficient vector search
PyMuPDF>=1.23.0

python-pptx>=0.6.21
striprtf>=0.0.26
PyPDF2>=2.10.0  # Fallback PDF processor


PyMuPDF>=1.24.0  # Or a specific version you need
python-docx>=1.1.0 # Or a specific version you need